--- selfdrive/qtui/.gitignore
@@ -1,3 +1,0 @@
-ui
-moc*
-*.o
