--- selfdrive/qtui/window.hpp
@@ -1,17 +1,0 @@
-#pragma once
-
-#include <QWidget>
-#include <QPushButton>
-#include <QVBoxLayout>
-#include <QHBoxLayout>
-#include <QSpacerItem>
-#include <QLabel>
-
-class MainWindow : public QWidget
-{
-  Q_OBJECT
-
-public:
-  explicit MainWindow(QWidget *parent = 0);
-
-};
