+++ selfdrive/ui/.gitignore
@@ -1,0 +1,1 @@
+moc_*
